Move pullRelevantPlotProps to plots.tsx
pullRelevantPlotProps builds PlotProps, which is defined in plots.tsx
-- but the function itself lived in comparison-panel.tsx, forcing
article-panel.tsx (the single-article view, not a comparison) to
import from the comparison UI component just to assemble plot data.

Pure move, no behavior change: comparison-panel.tsx now imports
pullRelevantPlotProps from plots.tsx instead of defining it,
article-panel.tsx imports it from plots.tsx directly.
