+++ react/src/components/comparison-panel.tsx
@@ -11,12 +11,10 @@ import { Navigator } from '../navigation/Navigator'
 import { colorFromCycle, useColors } from '../page_template/colors'
 import { rowExpandedKey, useSettings } from '../page_template/settings'
 import { groupYearKeys, StatGroupSettings } from '../page_template/statistic-settings'
-import { statParents, Year } from '../page_template/statistic-tree'
 import { PageTemplate } from '../page_template/template'
 import { compareArticleRows } from '../sorting'
 import { Universe, universeContext } from '../universe'
 import { mixWithBackground } from '../utils/color'
-import { assert } from '../utils/defensive'
 import { sanitize } from '../utils/paths'
 import { notWaiting, waiting } from '../utils/promiseStream'
 import { Article } from '../utils/protos'
@@ -30,7 +28,7 @@ import { computeNameSpecsWithGroups } from './article-panel'
 import { generateCSVDataForArticles, CSVExportData } from './csv-export'
 import { ArticleRow, isCongressionalRepresentativesMetadataRow, isNoValue } from './load-article'
 import { CommonMaplibreMap, PolygonFeatureCollection, polygonFeatureCollection, useZoomAllFeatures, defaultMapPadding, CustomAttributionControlComponent } from './map-common'
-import { PlotProps } from './plots'
+import { PlotProps, pullRelevantPlotProps } from './plots'
 import { createScreenshot, ScreencapElements, useScreenshotMode } from './screenshot'
 import { computeComparisonWidthColumns, computeMaxColumns, MaybeScroll } from './scrollable'
 import { SearchBox } from './search'
@@ -415,54 +413,6 @@ export function ComparisonPanel(props: {
             </TransposeContext.Provider>
         </universeContext.Provider>
     )
-}
-
-export function pullRelevantPlotProps(rows: ArticleRow[], statIndex: number, color: string, shortname: string, longname: string, sharedTypeOfAllArticles: string | undefined): PlotProps[] {
-    if (rows[statIndex].kind !== 'statistic' || rows[statIndex].extraStats.length === 0) {
-        return []
-    }
-    const sPs = rows.map(row => statParents.get(row.statpath)!).map((sP, i) => ({ sP, i }))
-    const byYear = new Map<Year, number[]>()
-    sPs.filter((
-        { sP, i }) => sP.group.id === sPs[statIndex].sP.group.id && rows[i].kind === 'statistic' && rows[i].extraStats.length > 0,
-    ).forEach(({ sP: { year }, i }) => {
-        assert(year !== null, 'Year should not be null for plot data')
-        byYear.set(year, [...(byYear.get(year) ?? []), i])
-    })
-    const bestSourceEach = Array.from(byYear.entries()).map(([, indices]) => {
-        if (indices.length === 1) {
-            return indices[0]
-        }
-        const sources = indices.map(i => sPs[i].sP.source)
-        const exactMatch = sources.findIndex(source => JSON.stringify(source) === JSON.stringify(sPs[statIndex].sP.source))
-        if (exactMatch !== -1) {
-            return indices[exactMatch]
-        }
-        const nullMatch = sources.findIndex(source => source === null)
-        if (nullMatch !== -1) {
-            return indices[nullMatch]
-        }
-        return indices[0]
-    })
-    const statpaths = bestSourceEach.map(i => sPs[i])
-    const overOne = statpaths.length > 1
-    if (overOne) {
-        statpaths.forEach(({ sP: { year } }) => {
-            assert(year !== null, 'Year should not be null for plot data')
-        })
-        assert(statpaths.length === new Set(statpaths.map(({ sP: { year } }) => year)).size, 'All statpaths for plot data should have unique years')
-    }
-    return statpaths.map(({ i: idx, sP: { year } }) => {
-        assert(year !== null, 'unreachable, we checked this already')
-        return {
-            ...rows[idx],
-            color,
-            shortname,
-            longname,
-            sharedTypeOfAllArticles,
-            subseriesName: year.toString(),
-        } satisfies PlotProps
-    })
 }
 
 function getHighlightIndex(rows: readonly ArticleRow[]): number | undefined {
